add cmake file
